+++ NativeTimelapseGenerator/console.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "worker-structs.h"
 void update_worker_stats(int worker_step, int count);
-void update_backups_stats(int backups_total, int backups_per_second, struct canvas_info current_info);
+void update_backups_stats(int backups_total, float backups_per_second, struct canvas_info current_info);
 void log_message(const char* format, ...);
 void* start_console(void* data);
 void stop_console();
